Move the MLX erf family into its own dispatch module
